Normalize the RMSE with the range and the mean in the min=max case
Training the model with the rewritten NRMSE in both cases lead to the loss function approaching the nan value. For now, the NRMSE is normalized by the range (max(signal)-min(signal)) and the min(signal)=max(signal) case is handled by normalizing the RMSE with the sample mean of the signal.
In any case, both ways of normalizing the metric seemed to be inefficient in removing the nan loss issue, which means that there are a few other options to try out:
1. using another gradient descent optimizer
2. using a smaller model