Save recordings as WAV files and update README.md
I just realised there was a small mistake in the README.md description of the project:
The model doesn't actually increase the sample rate (the downsampled audio is interpolated, so the low-res, high-res and super-res variants of the recording all have the same number of samples and consequently the same duration). Rather, it just learns how an audio signal (in this case a recording of a voice) is structured, therefore it is able to guess a realistic prediction of the frequencies that are missing from the recording.

The model seems to work quite well, but it is only saved as a checkpoint
instead of having its weights saved or being saved as a whole. That is
one thing that should be implemented soon.

The predict.py script takes some random track between indexes 100 and
200 (because the 0-99 range has been used for training). This awkward splitting
issue should be fairly easy to fix, because tfds.load(...) allows the
use of a parameter which splits the data automatically (by splitting all
of the dataset with this feature and regenerating the low-res/high-res
pairs of data, the model can be fully trained on the whole dataset, but
I'm not sure if that idea is worth implementing, because generating GB
of data for the task would also require even more time for re-training
the model).

The next step would be to implement the time-frequency model (which uses
both the time-domain and the frequency-domain representations of the
signal) and to read more papers about music super-resolution (which is
more or less the intended purpose of the project).

Also, just as a side-note: the code for the current version of the
project looks awful and uses no functions in a structured manner, so it might be a good idea to tidy up the code.