Add model with B=3 layers and 25ms chunks as inputs
The paper mentions an input of length 6000 in the high-resolution space. One thing that is unclear is how the output of the model mentioned in the paper should look like.
The B=3 layers model receives an input of 1200 samples (25ms, 48 kHz sample rate) and produces an output of 4800 samples (100ms, 48 kHz sample rate). We'll try to train this and evaluate it as such.
The only reason why this version of the model has 3 layers on both the downsampling and the upsampling path is that the output shapes of the intermediary layers do not match when concatenating results (due to the specific size of the input tensor which is 4800, there is some divisibility issue when applying convolutions, so no more layers can be used).
A larger input will be given in later versions of the model.