Fix nan loss issue related to the optimizer learning rate
The number of filters for each layer has been reduced to 32 currently, due to the former nan loss issue which appeared after a sudden extreme increase in the loss. That is why the learning rate is now 5e-4 instead of the default learning rate of Adam, which is 1e-3.
However, the early-stopper stopped the training after 10 epochs of no improvements in the validation loss.
The next step is to remove the early-stopper and train the model entirely for 100 epochs.