Finish training the 100-epoch model with a validation set of 1500 samples
The validation loss still seems to be much much lower than the training loss. The following article (https://stats.stackexchange.com/questions/187335/validation-error-less-than-training-error) reveals the possibility that the training set had many "hard cases to learn", hence why the validation loss is lower, but I'm not sure about this.
This might still be an unrepresentative validation dataset, as it is formed only of 5% of the whole available data. Perhaps the training/validation/testing data split needs to be changed.
Another cause for this might be the fluctuations created by the dropout regularization, because dropout is activated during training and deactivating during the validation phase.
The last resort option would be to increase the learning rate from 0.0005 to 0.00075.