Add interpolated version in the plots
The model clearly needs more training epochs or more filters in the convolutional layers (as the paper also mentions), because the input size is large and requires much more parameters to learn a pattern in the data.
One good idea would be to try training models on AWS Sagemaker to save time.