Add Colab training notebook with preprocessed dataset archive extraction
I ditched the idea of using AWS Sagemaker due to the costs and remembered that Colab was free and had free GPU training.
I had some issues when uploading the dataset, because I thought the dataset could be uploaded as a folder to Google Drive, but that was a mistake. Eventually, I zipped the folder, uploaded it to Drive and extracted it with the unzip command in Colab.