Add two models trained with 10^-4 learning rate and fix prediction script
The prediction script used to fail for recordings whose length was not divisible with BATCH_SIZE*SAMPLE_DIMENSION (16*4800).
To solve the issue, the remaining number of samples that the recording was missing has been replaced by zeroes and then the batch was fed to the model. Finally, the output's size has been cropped in such a way that the output has the same length as its high-resolution counterpart.