Finish training the 64-filters model
TODO: Generate plots for a fixed sample for all models (all the plots currently generated are each based on a random chunk picked from the dataset). This should show whether there is a significant difference between the models.
The spectrogram of the model containing 64 filters in each layer shows that the super-res version recovers some high frequencies, but it also generates some unwanted noise. Perhaps a good idea would be to perform an ablation study on the model and check how each of the layers contribute to the noise in the output.