Use different versions of NRMSE for the training/testing/validation dataset each with their corresponding IQR
It is difficult to compute the NRMSE with the IQR of each split set without creating separate functions for each, hence why there are NRMSE functions each with their corresponding interquartile range. The values for the IQR written in the constants.py file have been computed in the training_set_testing_set_distributions.ipynb

Now finally, the NRMSE values seem to be more realistic and not < .001.