Add the 3-layer model with the IQR-normalized RMSE
The loss plot still displays a much lower validation loss compared to the training loss. This might mean that the model overfits or that the validation set is not representative (perhaps increasing the size of the validation dataset might solve the issue). Another possibility could be that the Dropout regularization layer is turned off for the validation batches, hence why the validation loss is very small. Moreover, "Keras computes the average of the losses for each batch of the training set", as explained here: https://keras.io/getting_started/faq/#why-is-my-training-loss-much-higher-than-my-testing-loss