Add the trained 5-layer model with inception blocks
Oddly enough, the values of the training loss plot are closer to the validation loss values, but still slightly larger. The most plausible explanation is the dropout regularization, which is deactivated during the validation phases. (https://towardsdatascience.com/what-your-validation-loss-is-lower-than-your-training-loss-this-is-why-5e92e0b1747e)

The only remaining task is to run the testing script on the models
containing Inception modules/"multiscale convolutions".