Remove the dropout layer from the 5-layer Inception block model and retrain
We notice that the loss values seem to have improved for both the training and validation sets. However, there is a peculiar spike in the validation loss for the model with Inception blocks that do not contain a dropout layer. The following StackOverflow post delves deeper into the possibly negative effect of a dropout layer: https://stats.stackexchange.com/questions/299292/dropout-makes-performance-worse.